+++ src/constants/index.tsx
@@ -8,7 +8,7 @@ export const ST_PATH = '/stores';
 export const HO_PATH = '/how';
 
 export const SU_PATH = '/support/notice';
-export const SU_QA_PATH = '/support/qa';
+export const SU_QA_PATH = `${SU_PATH}/qa`;
 
 export const JO_PATH = '/join';
