+++ src/types/index.ts
@@ -1,5 +1,7 @@
 import NoticeList from "./notice.interface";
+import QaList from "./qa.interface";
 
 export type {
-    NoticeList
+    NoticeList,
+    QaList
 }
