+++ src/view/Support/style.css
@@ -1,102 +1,3 @@
 #support-wrapper {
     width: 1745px;
 }
-
-#notice-wrapper {
-    display: flex;
-    flex-direction: column;
-    gap: 30px;
-}
-
-#notice-wrapper .title {
-    text-align: center;
-    font-size: 30px;
-}
-
-#notice-wrapper .navigation {
-    display: flex;
-    justify-content: center;
-    gap: 60px;
-
-    font-size: 20px;
-    cursor: pointer;
-}
-
-#notice-wrapper .top-text {
-    font-size: 18px;
-}
-
-#notice-wrapper .table {
-    display: flex;
-    flex-direction: column;
-
-    border-top: 2px solid rgba(0, 0, 0, 1);
-    border-bottom: 2px solid rgba(0, 0, 0, 1);
-}
-
-#notice-wrapper .table .th {
-    display: flex;
-    justify-content: space-between;
-
-    border-bottom: 2px solid rgba(0, 0, 0, 1);
-
-    font-size: 18px;
-    font-weight: 700;
-
-    padding: 8px 50px;
-}
-
-#notice-wrapper .table .tr {
-    display: flex;
-    gap: 40px;
-
-    border-bottom: 1px solid rgba(0, 0, 0, 1);
-    padding: 8px 25px;
-}
-
-#notice-wrapper .table .tr .td-no {
-    width: 70px;
-    text-align: center;
-}
-
-#notice-wrapper .table .tr .td-title {
-    flex: 1;
-}
-
-#notice-wrapper .table .tr .td-date {
-    width: 130px;
-    text-align: center;
-}
-
-#notice-wrapper .bottom {
-    position: relative;
-    display: flex;
-    justify-content: center;
-}
-
-#notice-wrapper .search-box {
-    position: absolute;
-    right: 0;
-
-    display: flex;
-    justify-content: end;
-    align-items: center;
-    gap: 8px;
-}
-
-#notice-wrapper input {
-    padding: 7px 0px 7px 8px;
-    width: 145px;
-    border-radius: 5px;
-
-    outline: none;
-    border: 2px solid rgb(161, 161, 161);
-}
-
-#notice-wrapper .button {
-    padding: 7px 20px;
-    border-radius: 5px;
-
-    background-color: rgba(0, 0, 0, 1);
-    color: rgba(255, 255, 255, 1);
-}
